moving trace tests to a new package
