+++ src/tests/trace/TraceVisualizationTest.java
@@ -2,7 +2,7 @@
  * To change this template, choose Tools | Templates
  * and open the template in the editor.
  */
-package tests;
+package tests.trace;
 
 import gui.TraceVisualizer;
